Add some more known details to the MPK unpacker.
The Unknow0x06 field is the file version. Presumably there are version
0, version 1 and version 2 MPK files.
The version and number of files can be read as signed, since they're
cast to signed later anyway.
Version 0 and 1 files have smaller offset and size fields.
The Unknown value for each file is the compression flag. 0x00 is
uncompressed, 0x01 (or anything != 0x00?) is compressed.